requirements.txt
torch
transformers
datasets
triton
peft
pyyaml
scikit-learn